EAG/agent/requirements.txt
# Core data handling and analysis
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0

# Web scraping and API requests
requests>=2.31.0
beautifulsoup4>=4.12.0
requests-html>=0.10.0
aiohttp>=3.9.0

# News feed parsing
feedparser>=6.0.10
python-dateutil>=2.8.2

# API and environment handling
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# HTML parsing
lxml>=4.9.0

# Async support
asyncio>=3.4.3

# Time zone handling
pytz>=2023.3

# JSON handling
jsonschema>=4.20.0

# SSL support
certifi>=2023.11.17 